refactor: move Index from Xmi::Index to Xmi::Sparx::Index
Index only works with SparxRoot, so it belongs in the Sparx namespace.
Removed lib/xmi/index.rb, added autoload entry in sparx.rb, removed
require from xmi.rb.
